Migrated query related methods to the query building DAL
