+++ query-connector/src/app/(pages)/queryBuilding/page.test.tsx
@@ -1,11 +1,11 @@
 import QueryBuilding from "./page";
 import { render, screen, waitFor } from "@testing-library/react";
-import { DataContext } from "@/app/shared/DataProvider";
 import {
   getConditionsData,
   getCustomQueries,
 } from "../../shared/database-service";
 import { conditionIdToNameMap, DEFAULT_QUERIES } from "./fixtures";
+import { RootProviderMock } from "@/app/tests/unit/setup";
 
 jest.mock(".../../../shared/database-service", () => ({
   getCustomQueries: jest.fn(),
@@ -16,22 +16,10 @@ describe("tests the query building steps", () => {
   it("renders the empty state", async () => {
     (getCustomQueries as jest.Mock).mockResolvedValue([]);
 
-    const mockSetData = jest.fn();
-    const mockSetCurrentPage = jest.fn();
-    const mockSetToatConfig = jest.fn();
-    const mockContextValue = {
-      data: undefined,
-      setData: mockSetData,
-      currentPage: "/queryBuilding",
-      setCurrentPage: mockSetCurrentPage,
-      toastConfig: {},
-      setToastConfig: mockSetToatConfig,
-    };
-
     render(
-      <DataContext.Provider value={mockContextValue}>
+      <RootProviderMock currentPage="/queryBuilding">
         <QueryBuilding />
-      </DataContext.Provider>,
+      </RootProviderMock>,
     );
 
     await waitFor(() => {
@@ -47,22 +35,10 @@ describe("tests the query building steps", () => {
       conditionIdToNameMap,
     });
 
-    const mockSetData = jest.fn();
-    const mockSetCurrentPage = jest.fn();
-    const mockSetToatConfig = jest.fn();
-    const mockContextValue = {
-      data: DEFAULT_QUERIES,
-      setData: mockSetData,
-      currentPage: "/queryBuilding",
-      setCurrentPage: mockSetCurrentPage,
-      toastConfig: {},
-      setToastConfig: mockSetToatConfig,
-    };
-
     render(
-      <DataContext.Provider value={mockContextValue}>
+      <RootProviderMock currentPage="/queryBuilding">
         <QueryBuilding />
-      </DataContext.Provider>,
+      </RootProviderMock>,
     );
 
     const expectedQueryNames = DEFAULT_QUERIES.map((q) => q.query_name);
